[ci] Switch to use cargo-deny GitHub action
This also upgrades from cargo-deny 0.4.0 to 0.4.2.

This uses master version of the action, but we should switch soon to a stable v1 of it. But helpful to use master to start with to help stabilize the action.

Also fixed that the publish and release steps didn't require that cargo-deny was successful.